custom alert added

+++ src/pages/ManageInventories/ManageInventoriesData/ManageInventoriesData.js
@@ -3,20 +3,53 @@ import { faPen, faTrashCan } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { toast, ToastContainer } from 'react-toastify';
 import { Link } from 'react-router-dom';
+import Swal from 'sweetalert2';
 const ManageInventoriesData = ({ car }) => {
     const { name, img, _id, price, qty, des, supplier, brand } = car;
 
-
+    const swalWithBootstrapButtons = Swal.mixin({
+        customClass: {
+            confirmButton: 'btn btn-success',
+            cancelButton: 'btn btn-danger'
+        },
+        buttonsStyling: false
+    })
     const deleteCar = () => {
-        const url = `http://localhost:5000/car?_id=${_id}`
-        fetch(url, {
-            method: 'DELETE'
+        swalWithBootstrapButtons.fire({
+            title: 'Are you sure?',
+            text: `Want to delete ${name} ?`,
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonText: 'Yes, delete it!',
+            cancelButtonText: 'No, cancel!',
+            reverseButtons: true
+        }).then((result) => {
+            if (result.isConfirmed) {
+                const url = `http://localhost:5000/car?_id=${_id}`
+                fetch(url, {
+                    method: 'DELETE'
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        swalWithBootstrapButtons.fire(
+                            'Deleted!',
+                            `${name} has been successfully deleted.`,
+                            'success'
+                        )
+                    })
+
+            } else if (
+                /* Read more about handling dismissals below */
+                result.dismiss === Swal.DismissReason.cancel
+            ) {
+                swalWithBootstrapButtons.fire(
+                    'Cancelled',
+                    `${name} Item not deleted`,
+                    'error'
+                )
+            }
         })
-            .then(res => res.json())
-            .then(data => {
-                console.log('Delete Done');
-                toast.success('Delete Done')
-            })
+
     }
 
     return (
